#1 partial fix for oneOf always using interfaces
This is only a partial fix as it still relies on being able to turn referenced models into interfaces, and I’m not sure how a Java generator will deal with this.

This also changes the anyOf behaviour to adopt the old oneOf-without-discriminator behaviour, which is to make a concrete class the model.