Remove usages of toCodegenModel and move to using toCodegenSchema
This acknowledges all the things that are _actually_ schemas and not always a model (which is a specialisation of a schema). We can now handle oneOf with a schema that isn’t a model without having to pretend it’s okay in toCodegenModel.

Note that there is an unpleasant workaround in toCodegenModel for the aforementioned case, that will be removed once I resolve the use of terms like “implements” that are Java terms that don’t apply to this scenario!