refactor: CodegenSchemaUse
It started with required, which was in CodegenSchemaInfo. But it was related to where the schema was used, not to the schema itself. So now CodegenSchemaUse represents the usage of a schema, so it contains required and the required is applied to the schema (which might change it, say from long to Long). Also the examples and default value are now in the CodegenSchemaUse as they change according to the type.