Add initialValue to CodegenProperty
Properties previously always had a defaultValue, because CodegenSchema did. Even if the specification didn’t include a default. We used this to set the initial value in the model object in Java for convenience. The initialValue now performs that function of always having a value, whether there was an example or not.

This corrects the new issue of things like request bodies getting default values when there were none specified.