patch-package: @apidevtools/json-schema-ref-parser to fix relative refs in external docs
When the parser parses an external document it leaves relative refs in the external document as they are, but that means when we later come to resolve them we can’t as they look like they’re relative to the original document. The patch corrects the refs so they include the external document.