core: force models referenced by a discriminator to be loaded
We might not load models that are referenced by a discriminator if they appear in an external file. Now we explicitly load them when we see them. I think this only affects the parent class, subclass style of discriminator as if you reference the parent class you wouldn’t know about the subclasses _unless_ you looked at the mappings.

This means that external references are flawed as if there weren’t any mappings, which is totally fine, we wouldn’t discover the subclasses! TODO I guess.