default value / initial value
Previously this was treated a bit uncertainly. Now it is firmed up and the API to generators is changed so that generators can provide the right information; do we need a default value to apply to a property? Use defaultValue. Do we need to know what initial value we’d like for a property? Use initialValue. These are different questions. We also now support returning null to say there is no initial value.