native type usage transformer
We previously had this weird situation where we tried to use `required` to determine whether a native type was primitive or object (java), but we needed to determine when we resolved the schema, not where we used the schema, and nullable could change! And required wasn’t really known. So we had a hack in schema/index to re-do the native type…

Now we introduce a nativeTypeUsageTransformer so we can transform a native type according to what we learn when we try to use it.