component: we need to store CodegenSchemaUsage to get the right nativeType
At some point I think I changed to storing the CodegenSchema instead of CodegenSchemaUsage as I think I saw a lot of componentSchema.schema in templates. Doesn’t seem to be a thing anymore. The CodegenSchemaUsage usaully contains what you want—the nativeType—and we must use that native type _not_ the one from the underlying schema as the native type in the _usage_ reflects nullabilty etc.