objects: revert to using raw property name in properties collection
Regression from 86e1cff4890861b3a472849c6976550c299550d4. We need to use the original property name from the spec, as things refer to it (such as the upcoming encoding work), therefore we need to ensure uniqueness of the name _inside_ the property, knowing it might not be the same as the key the property is under.