array/map: remove the purpose parameter
See e9c2c2f62944a26e139c96e26019e45ac700a347. This was used when we supported having objects that had arrays and maps as parents. This is no longer a thing (I don't know when it was removed) so we're taking the leap and removing the purpose as well, as it's now defunct.