composition: support multiple strategies for allOf, anyOf, oneOf; explicit interface support
This is the major overhaul of how we handle allOf, anyOf and oneOf. We now support multiple ways of translating these concepts to a native language, giving the generator options to choose which approach best suits their language, and providing a foundation to support additional strategies in future.

Interfaces are now a separate schema type, rather than being stuffed into CodegenObjectSchema. I hope the concept of an interface can be adapted to any language. Inheritance is still supported on objects.

Wrappers are introduced to support primitive types in compositions.