nativeType: remove nulls and recursion, fix equals implementation
We don't need to express that we don't support a given form of the native type. That's not how we represent whether we can do something or not. We didn't check it anywhere.

The recursion in componentType was a pain for logging and for the implementation of equals.

The equals implementation now hopefully works correctly!